requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
Django==2.2
django-allauth==0.41.0
django-crispy-forms==1.8.1
django-tinymce4-lite==1.7.5
idna==2.8
isort==4.3.21
jsmin==2.2.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==6.2.0
pycodestyle==2.5.0
pylint==2.4.4
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.8
wrapt==1.11.2
